requirements.txt
flask==2.2.*
psycopg2-binary==2.9.*